Move reviewVoteUpdate.ts from lib->server
