+++ packages/lesswrong/server/manualMigrations/2022-12-13-updateReviewVotes.ts
@@ -1,4 +1,4 @@
-import { updateReviewVoteTotals } from '../../lib/reviewVoteUpdate';
+import { updateReviewVoteTotals } from '../reviewVoteUpdate';
 import { registerMigration } from './migrationUtils';
